+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixes 🍕
 
 - Fixed path of `use_color_mode` cookie
+- `ElementMaybeSignal` and `ElementsMaybeSignal` are now properly `Clone` and `Copy`
 
 ## [0.15.1] - 2024-12-31